+++ app/(main)/(inside)/(screens)/(activity)/applications.tsx
