+++ app/(main)/(inside)/(screens)/(activity)/contracts.tsx
