+++ app/(main)/(inside)/(screens)/(profile)/bank-shipping-address.tsx
