+++ app/(main)/(inside)/(screens)/(profile)/notifications.tsx
