+++ app/(main)/(inside)/(screens)/(profile)/preferences.tsx
