+++ app/(main)/(inside)/(screens)/apply-now/[pageID].tsx
