+++ app/(main)/(inside)/(screens)/apply-now/[pageID].web.tsx
