+++ app/(main)/(inside)/(screens)/apply-now/_layout.tsx
