+++ app/(main)/(inside)/(screens)/apply-now/preview.tsx
