+++ app/(main)/(inside)/(screens)/apply-now/question.tsx
