+++ app/(main)/(inside)/(screens)/apply-now/quotation.tsx
