+++ app/(main)/(inside)/(screens)/collaboration-details/[pageID].tsx
