+++ app/(main)/(inside)/(screens)/contract-details/[pageID].tsx
