+++ app/(main)/(inside)/(screens)/edit-application/[pageID].web.tsx
