+++ app/(main)/(inside)/(tabs)/_layout.tsx
