+++ app/(main)/(inside)/(tabs)/collaborations.tsx
