+++ app/(main)/(inside)/(tabs)/influencers.tsx
